requirements.txt
altair
annotated-types
anthropic
anyio
appnope
asttokens
attrs
beautifulsoup4
blinker
bs4
cachetools
certifi
charset-normalizer
click
comm
debugpy
decorator
distro
dnspython
duckduckgo-search
exceptiongroup
executing
feedparser
filelock
fsspec
gitdb
gitpython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
groq
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
hugchat
huggingface-hub
idna
importlib-metadata
ipykernel
ipython
ipython-pygments-lexers
jedi
jinja2
jiter
jsonschema
jsonschema-specifications
jupyter-client
jupyter-core
lxml
markdown-it-py
markupsafe
matplotlib-inline
mdurl
mpmath
narwhals
nest-asyncio
networkx
numpy
ollama
openai
orjson
packaging
pandas
parso
pexpect
pickleshare
pillow
platformdirs
primp
prompt-toolkit
proto-plus
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pydeck
pygments
pymongo
pyparsing
pyreqwest-impersonate
python-dateutil
python-dotenv
pytz
pyyaml
pyzmq
referencing
regex
requests
requests-toolbelt
rich
rpds-py
rsa
safetensors
setuptools
sgmllib3k
six
smmap
sniffio
soupsieve
stack-data
streamlit
style
sympy
tavily-python
tenacity
tiktoken
tokenizers
toml
toolz
torch
tornado
tqdm
traitlets
transformers
typing-extensions
tzdata
update
uritemplate
urllib3
validators
watchdog
wcwidth
zipp
